@@ -126,58 +126,58 @@
         "prestashop/statssales": "^2",
         "prestashop/statssearch": "^2",
         "prestashop/statsstock": "^2",
-        "prestashop/translationtools-bundle": "^6",
-        "sensio/framework-extra-bundle": "^5.1",
+        "prestashop/translationtools-bundle": "dev-master",
+        "sensio/framework-extra-bundle": "^6.2",
         "smarty/smarty": "^4.3.1",
         "soundasleep/html2text": "^2.1.0",
-        "symfony/asset": "5.4.*",
-        "symfony/cache": "5.4.*",
-        "symfony/config": "5.4.*",
-        "symfony/console": "5.4.*",
-        "symfony/dependency-injection": "5.4.*",
-        "symfony/doctrine-messenger": "5.4.*",
-        "symfony/dom-crawler": "5.4.*",
-        "symfony/dotenv": "5.4.*",
-        "symfony/error-handler": "5.4.*",
-        "symfony/event-dispatcher": "5.4.*",
-        "symfony/expression-language": "5.4.*",
-        "symfony/filesystem": "5.4.*",
-        "symfony/finder": "5.4.*",
-        "symfony/form": "5.4.*",
-        "symfony/http-client": "5.4.*",
-        "symfony/http-foundation": "5.4.*",
-        "symfony/http-kernel": "5.4.*",
-        "symfony/intl": "5.4.*",
-        "symfony/lock": "5.4.*",
-        "symfony/mailer": "5.4.*",
-        "symfony/messenger": "5.4.*",
-        "symfony/mime": "5.4.*",
+        "symfony/asset": "6.0.*",
+        "symfony/cache": "6.0.*",
+        "symfony/config": "6.0.*",
+        "symfony/console": "6.0.*",
+        "symfony/dependency-injection": "6.0.*",
+        "symfony/doctrine-bridge": "6.0.*",
+        "symfony/doctrine-messenger": "6.0.*",
+        "symfony/dom-crawler": "6.0.*",
+        "symfony/dotenv": "6.0.*",
+        "symfony/error-handler": "6.0.*",
+        "symfony/event-dispatcher": "6.0.*",
+        "symfony/expression-language": "6.0.*",
+        "symfony/filesystem": "6.0.*",
+        "symfony/finder": "6.0.*",
+        "symfony/form": "6.0.*",
+        "symfony/http-client": "6.0.*",
+        "symfony/http-foundation": "6.0.*",
+        "symfony/http-kernel": "6.0.*",
+        "symfony/intl": "6.0.*",
+        "symfony/lock": "6.0.*",
+        "symfony/mailer": "6.0.*",
+        "symfony/messenger": "6.0.*",
+        "symfony/mime": "6.0.*",
         "symfony/monolog-bundle": "^3.7.0",
-        "symfony/options-resolver": "5.4.*",
-        "symfony/password-hasher": "5.4.*",
+        "symfony/options-resolver": "6.0.*",
+        "symfony/password-hasher": "6.0.*",
         "symfony/polyfill-php73": "^1.10",
-        "symfony/process": "5.4.*",
-        "symfony/property-access": "5.4.*",
-        "symfony/property-info": "5.4.*",
-        "symfony/proxy-manager-bridge": "5.4.*",
+        "symfony/process": "6.0.*",
+        "symfony/property-access": "6.0.*",
+        "symfony/property-info": "6.0.*",
+        "symfony/proxy-manager-bridge": "6.0.*",
         "symfony/psr-http-message-bridge": "^2.1",
-        "symfony/routing": "5.4.*",
-        "symfony/security-bundle": "5.4.*",
-        "symfony/security-core": "5.4.*",
-        "symfony/security-csrf": "5.4.*",
-        "symfony/security-guard": "5.4.*",
-        "symfony/security-http": "5.4.*",
-        "symfony/serializer": "5.4.*",
-        "symfony/string": "5.4.*",
-        "symfony/templating": "5.4.*",
-        "symfony/translation": "5.4.*",
-        "symfony/twig-bundle": "5.4.*",
+        "symfony/routing": "6.0.*",
+        "symfony/security-bundle": "6.0.*",
+        "symfony/security-core": "6.0.*",
+        "symfony/security-csrf": "6.0.*",
+        "symfony/security-http": "6.0.*",
+        "symfony/serializer": "6.0.*",
+        "symfony/string": "6.0.*",
+        "symfony/templating": "6.0.*",
+        "symfony/translation": "6.0.*",
+        "symfony/twig-bundle": "6.0.*",
         "symfony/ux-twig-component": "^2.9",
-        "symfony/validator": "5.4.*",
-        "symfony/var-dumper": "5.4.*",
-        "symfony/var-exporter": "5.4.*",
-        "symfony/web-link": "5.4.*",
-        "symfony/yaml": "5.4.*",
+        "symfony/validator": "6.0.*",
+        "symfony/var-dumper": "6.0.*",
+        "symfony/var-exporter": "6.0.*",
+        "symfony/web-link": "6.0.*",
+        "symfony/yaml": "6.0.*",
         "tecnickcom/tcpdf": "^6.2.12",
         "tijsverkoyen/css-to-inline-styles": "^2.2",
         "twig/extra-bundle": "^3.6",
@@ -195,12 +195,12 @@
         "phpunit/phpunit": "~9.6.7",
         "rector/rector": "^0.15.17",
         "spaze/phpstan-disallowed-calls": "^2.10",
-        "symfony/browser-kit": "5.4.*",
-        "symfony/css-selector": "5.4.*",
-        "symfony/debug-bundle": "5.4.*",
+        "symfony/browser-kit": "6.0.*",
+        "symfony/css-selector": "6.0.*",
+        "symfony/debug-bundle": "6.0.*",
         "symfony/phpunit-bridge": "6.2.7",
-        "symfony/stopwatch": "5.4.*",
-        "symfony/web-profiler-bundle": "5.4.*"
+        "symfony/stopwatch": "6.0.*",
+        "symfony/web-profiler-bundle": "6.0.*"
     },
     "config": {
         "allow-plugins": {
