@@ -35,7 +35,7 @@ use Psr\Log\LoggerInterface;
  */
 class LegacyLogger implements LoggerInterface
 {
-    public function emergency($message, array $context = [])
+    public function emergency($message, array $context = []): void
     {
         $this->log(Logger::EMERGENCY, $message, $context);
     }
@@ -49,7 +49,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function alert($message, array $context = [])
+    public function alert($message, array $context = []): void
     {
         $this->log(Logger::ALERT, $message, $context);
     }
@@ -62,7 +62,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function critical($message, array $context = [])
+    public function critical($message, array $context = []): void
     {
         $this->log(Logger::CRITICAL, $message, $context);
     }
@@ -74,7 +74,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function error($message, array $context = [])
+    public function error($message, array $context = []): void
     {
         $this->log(Logger::ERROR, $message, $context);
     }
@@ -88,7 +88,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function warning($message, array $context = [])
+    public function warning($message, array $context = []): void
     {
         $this->log(Logger::WARNING, $message, $context);
     }
@@ -99,7 +99,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function notice($message, array $context = [])
+    public function notice($message, array $context = []): void
     {
         $this->log(Logger::NOTICE, $message, $context);
     }
@@ -112,7 +112,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function info($message, array $context = [])
+    public function info($message, array $context = []): void
     {
         $this->log(Logger::INFO, $message, $context);
     }
@@ -123,7 +123,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function debug($message, array $context = [])
+    public function debug($message, array $context = []): void
     {
         $this->log(Logger::DEBUG, $message, $context);
     }
@@ -135,7 +135,7 @@ class LegacyLogger implements LoggerInterface
      * @param string $message
      * @param array $context
      */
-    public function log($level, $message, array $context = [])
+    public function log($level, $message, array $context = []): void
     {
         switch ($level) {
             case Logger::EMERGENCY:
