@@ -38,11 +38,12 @@ use PrestaShop\PrestaShop\Core\Security\Permission;
 use PrestaShopBundle\Translation\TranslatorInterface;
 use Psr\Container\ContainerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
+use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Contracts\Service\ServiceProviderInterface;
 
 /**
  * Extends The Symfony framework bundle controller to add common functions for PrestaShop needs.
@@ -50,30 +51,52 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
  * @deprecated since 9.0 to be removed in future versions (10+ at least, when it will not be used anymore),
  * should stop using it in favor of PrestaShopAdminController.
  */
-class FrameworkBundleAdminController extends AbstractController implements ContainerAwareInterface
+class FrameworkBundleAdminController extends AbstractController
 {
     /**
      * @deprecated since 9.0
      */
     public const PRESTASHOP_CORE_CONTROLLERS_TAG = 'prestashop.core.controllers';
 
-    /**
-     * Override to make this compatible with the ContainerAware signature, content should be the same as in the abstract.
-     * Do not override this neither use this, it will be removed in next versions. This overridden method
-     * along with the ContainerAwareInterface was added to skip the error sent by Symfony\Bundle\FrameworkBundle\Controller\ControllerResolver
-     * that forces the controllers extending AbstractController to be defined as service subscriber.
-     *
-     * This method allows us to keep controllers based on FrameworkBundleAdminController from being
-     * adapted. However, the core and the modules should stop using it in favor of PrestaShopAdminController
-     *
-     * @internal
-     */
-    public function setContainer(ContainerInterface $container = null): ?ContainerInterface
-    {
-        $previous = $this->container;
-        $this->container = $container;
+    protected ?ServiceProviderInterface $controllerContainer = null;
 
-        return $previous;
+    protected ?Container $globalContainer = null;
+
+    /**
+     * This method is completely hacky, we count on the fact that it is going to be used to inject the controller's dedicated
+     * minified controller (thanks to the @required annotation, autowiring and AbstractController parent class), this allows us
+     * to store the controller container in a dedicated field.
+     *
+     * On a second call, made by Symfony\Bundle\FrameworkBundle\Controller\ControllerResolver, this setter is called with the
+     * global container (mainly to cehck the current value actually), so we use the occasion to store the global container.
+     *
+     * The real container should be the controller one, but it doesn't contain all the public services we need that are in
+     * the global container, so we keep a reference on both containers so that the get and has methods can try fallback on
+     * both of them.
+     *
+     * This is quite ugly, but it prevents refactoring all the controllers (from both core and modules controllers), it is only
+     * done on controllers that extend this class which should not be used anymore and be replaced by PrestaShopAdminController
+     * controller by controller along with a refacto to do proper dependency injection.
+     *
+     * @param ContainerInterface $container
+     *
+     * @return ContainerInterface|null
+     *
+     * Note: this annotation is a MUST-HAVE, we have to keep it
+     *
+     * @required
+     */
+    public function setContainer(ContainerInterface $container): ?ContainerInterface
+    {
+        $return = parent::setContainer($container);
+        if ($container instanceof ServiceProviderInterface) {
+            $this->controllerContainer = $container;
+        }
+        if ($container instanceof Container) {
+            $this->globalContainer = $container;
+        }
+
+        return $return;
     }
 
     /**
@@ -85,6 +108,15 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function has(string $id): bool
     {
+        trigger_deprecation('prestashop/prestashop', '9.0', 'Method "%s()" is deprecated, use method or constructor injection in your controller instead.', __METHOD__);
+
+        if ($this->controllerContainer && $this->controllerContainer->has($id)) {
+            return true;
+        }
+        if ($this->globalContainer && $this->globalContainer->has($id)) {
+            return true;
+        }
+
         return $this->container->has($id);
     }
 
@@ -97,6 +129,35 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function get(string $id): object
     {
+        trigger_deprecation('prestashop/prestashop', '9.0', 'Method "%s()" is deprecated, use method or constructor injection in your controller instead.', __METHOD__);
+
+        return $this->doGet($id);
+    }
+
+    /**
+     * This special get method tries to get a service either in the controller custom-made container (that contains
+     * the most regular aliases to private services like twig, security, ...) and if it doesn't find it it tries to
+     * get it from the global container (that contains all public services).
+     *
+     * This is completely going around the framework, we do this to allow this class to behave as it used to without
+     * having to refactor the controller completely with proper dependncy injection, but it's a temporary solution that
+     * will disappear with this fix.
+     *
+     * @param string $id
+     *
+     * @throws \Psr\Container\ContainerExceptionInterface
+     * @throws \Psr\Container\NotFoundExceptionInterface
+     */
+    protected function doGet(string $id): object
+    {
+        if ($this->controllerContainer && $this->controllerContainer->has($id)) {
+            return $this->controllerContainer->get($id);
+        }
+
+        if ($this->globalContainer && $this->globalContainer->has($id)) {
+            return $this->globalContainer->get($id);
+        }
+
         return $this->container->get($id);
     }
 
@@ -109,7 +170,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function getDoctrine(): ManagerRegistry
     {
-        return $this->container->get('doctrine');
+        return $this->get('doctrine');
     }
 
     /**
@@ -122,7 +183,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function getConfiguration(): ShopConfigurationInterface
     {
-        return $this->container->get('prestashop.adapter.legacy.configuration');
+        return $this->get('prestashop.adapter.legacy.configuration');
     }
 
     /**
@@ -186,7 +247,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function dispatchHook($hookName, array $parameters)
     {
-        $this->container->get('prestashop.core.hook.dispatcher')->dispatchWithParameters($hookName, $parameters);
+        $this->get('prestashop.core.hook.dispatcher')->dispatchWithParameters($hookName, $parameters);
     }
 
     /**
@@ -203,7 +264,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function renderHook($hookName, array $parameters)
     {
-        return $this->container->get('prestashop.core.hook.dispatcher')->renderForParameters($hookName, $parameters)->getContent();
+        return $this->get('prestashop.core.hook.dispatcher')->renderForParameters($hookName, $parameters)->getContent();
     }
 
     /**
@@ -225,7 +286,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
         $iso = (string) $legacyContext->getEmployeeLanguageIso();
 
         $url = $this->generateUrl('admin_common_sidebar', [
-            'url' => $this->container->get(Documentation::class)->generateLink($section, $iso),
+            'url' => $this->get(Documentation::class)->generateLink($section, $iso),
             'title' => $title,
         ]);
 
@@ -281,7 +342,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function langToLocale($lang)
     {
-        return $this->container->get('prestashop.service.translation')->langToLocale($lang);
+        return $this->get('prestashop.service.translation')->langToLocale($lang);
     }
 
     /**
@@ -366,7 +427,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function redirectToDefaultPage()
     {
-        $legacyContext = $this->container->get('prestashop.adapter.legacy.context');
+        $legacyContext = $this->get('prestashop.adapter.legacy.context');
         $defaultTab = $legacyContext->getDefaultEmployeeTab();
 
         return $this->redirect($legacyContext->getAdminLink($defaultTab));
@@ -434,7 +495,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function getFallbackErrorMessage($type, $code, $message = '')
     {
-        $isDebug = $this->container->get('kernel')->isDebug();
+        $isDebug = $this->get('kernel')->isDebug();
         if ($isDebug && !empty($message)) {
             return $this->trans(
                 'An unexpected error occurred. [%type% code %code%]: %message%',
@@ -468,7 +529,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function getAdminLink($controller, array $params, $withToken = true)
     {
-        return $this->container->get('prestashop.adapter.legacy.context')->getAdminLink($controller, $withToken, $params);
+        return $this->get('prestashop.adapter.legacy.context')->getAdminLink($controller, $withToken, $params);
     }
 
     /**
@@ -480,7 +541,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function presentGrid(GridInterface $grid)
     {
-        return $this->container->get('prestashop.core.grid.presenter.grid_presenter')->present($grid);
+        return $this->get('prestashop.core.grid.presenter.grid_presenter')->present($grid);
     }
 
     /**
@@ -490,7 +551,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function getCommandBus()
     {
-        return $this->container->get('prestashop.core.command_bus');
+        return $this->get('prestashop.core.command_bus');
     }
 
     /**
@@ -500,7 +561,7 @@ class FrameworkBundleAdminController extends AbstractController implements Conta
      */
     protected function getQueryBus()
     {
-        return $this->container->get('prestashop.core.query_bus');
+        return $this->get('prestashop.core.query_bus');
     }
 
     /**
