@@ -55,20 +55,18 @@ class NullDispatcher implements EventDispatcherInterface, HookDispatcherInterfac
 
     /**
      * @param null $eventName
-     *
-     * @return array|void
      */
-    public function getListeners($eventName = null)
+    public function getListeners($eventName = null): array
     {
+        return [];
     }
 
     /**
      * @param null $eventName
-     *
-     * @return bool|void
      */
-    public function hasListeners($eventName = null)
+    public function hasListeners($eventName = null): bool
     {
+        return false;
     }
 
     public function removeListener($eventName, $listener)
@@ -82,11 +80,10 @@ class NullDispatcher implements EventDispatcherInterface, HookDispatcherInterfac
     /**
      * @param string $eventName
      * @param callable $listener
-     *
-     * @return int|void|null
      */
-    public function getListenerPriority($eventName, $listener)
+    public function getListenerPriority($eventName, $listener): ?int
     {
+        return null;
     }
 
     public function dispatchHook(HookInterface $hook)
