@@ -68,7 +68,7 @@ class IconButtonType extends ButtonType
     /**
      * {@inheritDoc}
      */
-    public function getParent()
+    public function getParent(): ?string
     {
         return ButtonType::class;
     }
@@ -76,7 +76,7 @@ class IconButtonType extends ButtonType
     /**
      * {@inheritdoc}
      */
-    public function getBlockPrefix()
+    public function getBlockPrefix(): string
     {
         return 'icon_button';
     }
