@@ -28,24 +28,18 @@ declare(strict_types=1);
 
 namespace Tests\Integration\Utility;
 
-use PrestaShop\PrestaShop\Adapter\LegacyContext;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
 
 /**
  * Admin Middleware security
+ *
+ * The listener behavior was removed in the tests in favor of the LoginTrait trait,
+ * to avoid necessarily having to have an employee in the context,
+ * which could have side effects on some test scenarios and also follow the
+ * official symfony method for authentication in a test environment.
  */
 class AdapterSecurityAdmin
 {
-    public function __construct(
-        private readonly LegacyContext $context,
-        private readonly TokenStorageInterface $securityTokenStorage,
-        private readonly UserProviderInterface $userProvider
-    ) {
-    }
-
     /**
      * Aims to authenticate the employee present in the context, only on routes not concerning the API
      *
@@ -55,17 +49,5 @@ class AdapterSecurityAdmin
      */
     public function onKernelRequest(RequestEvent $event): void
     {
-        $actualFirewall = $event->getRequest()->attributes->get('_firewall_context');
-
-        if (null !== $actualFirewall && (str_ends_with($actualFirewall, 'api_token') || str_ends_with($actualFirewall, 'api'))) {
-            return;
-        }
-        $employee = $this->context->getContext()->employee;
-
-        if (null !== $employee && null !== $employee->email) {
-            $user = $this->userProvider->loadUserByIdentifier($employee->email);
-            $token = new UsernamePasswordToken($user, 'admin', $user->getRoles());
-            $this->securityTokenStorage->setToken($token);
-        }
     }
 }
